pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.12.15",
    "asyncssh>=2.21.0",
    "click>=8.2.1",
    "fastapi>=0.116.1",
    "pydantic>=2.11.7",
    "pyyaml>=6.0.2",
    "uvicorn>=0.35.0",
]
